Renaming Protocol to Node and moving Crypto.Bloom to Node.Bloom
