+++ Network/Haskoin/Internals.hs
@@ -14,9 +14,9 @@ module Network.Haskoin.Internals
 , module Network.Haskoin.Crypto.NormalizedKeys
 , module Network.Haskoin.Crypto.ECDSA
 , module Network.Haskoin.Crypto.Mnemonic
-, module Network.Haskoin.Crypto.Bloom
-, module Network.Haskoin.Protocol.Message
-, module Network.Haskoin.Protocol.Types
+, module Network.Haskoin.Node.Types
+, module Network.Haskoin.Node.Message
+, module Network.Haskoin.Node.Bloom
 , module Network.Haskoin.Script.Types
 , module Network.Haskoin.Script.Parser
 , module Network.Haskoin.Script.SigHash
@@ -38,9 +38,9 @@ import Network.Haskoin.Crypto.ExtendedKeys
 import Network.Haskoin.Crypto.NormalizedKeys
 import Network.Haskoin.Crypto.ECDSA
 import Network.Haskoin.Crypto.Mnemonic
-import Network.Haskoin.Crypto.Bloom
-import Network.Haskoin.Protocol.Message
-import Network.Haskoin.Protocol.Types
+import Network.Haskoin.Node.Types
+import Network.Haskoin.Node.Message
+import Network.Haskoin.Node.Bloom
 import Network.Haskoin.Script.Types
 import Network.Haskoin.Script.Parser
 import Network.Haskoin.Script.SigHash
@@ -49,3 +49,4 @@ import Network.Haskoin.Transaction.Builder
 import Network.Haskoin.Transaction.Types
 import Network.Haskoin.Block.Types
 import Network.Haskoin.Block.Merkle
+
