+++ Network/Haskoin/Test/Crypto.hs
@@ -24,10 +24,6 @@ module Network.Haskoin.Test.Crypto
 , ArbitraryAccPubKey(..)
 , ArbitraryAddrPrvKey(..)
 , ArbitraryAddrPubKey(..)
-, ArbitraryBloomFlags(..)
-, ArbitraryBloomFilter(..)
-, ArbitraryFilterLoad(..)
-, ArbitraryFilterAdd(..)
 ) where
 
 import Test.QuickCheck
@@ -45,7 +41,6 @@ import Network.Haskoin.Crypto.Base58
 import Network.Haskoin.Crypto.Curve
 import Network.Haskoin.Crypto.ExtendedKeys
 import Network.Haskoin.Crypto.NormalizedKeys
-import Network.Haskoin.Crypto.Bloom
 
 -- | Arbitrary strict ByteString
 data ArbitraryByteString = ArbitraryByteString BS.ByteString
@@ -283,46 +278,4 @@ instance Arbitrary ArbitraryAddrPubKey where
         ArbitraryAddrPrvKey m k a <- arbitrary
         let p = AddrPubKey $ deriveXPubKey $ getAddrPrvKey a
         return $ ArbitraryAddrPubKey m k a p
-
--- | Arbitrary bloom filter flags
-data ArbitraryBloomFlags = ArbitraryBloomFlags BloomFlags
-    deriving (Eq, Show, Read)
-
-instance Arbitrary ArbitraryBloomFlags where
-    arbitrary = ArbitraryBloomFlags <$> elements 
-        [ BloomUpdateNone
-        , BloomUpdateAll
-        , BloomUpdateP2PubKeyOnly
-        ]
-
--- | Arbitrary bloom filter with its corresponding number of elements
--- and false positive rate.
-data ArbitraryBloomFilter = ArbitraryBloomFilter Int Double BloomFilter
-    deriving (Eq, Show, Read)
-
-instance Arbitrary ArbitraryBloomFilter where
-    arbitrary = do
-        n     <- choose (0,100000)
-        fp    <- choose (1e-8,1)
-        tweak <- arbitrary
-        ArbitraryBloomFlags fl <- arbitrary
-        return $ ArbitraryBloomFilter n fp $ bloomCreate n fp tweak fl
-
--- | Arbitrary FilterLoad
-data ArbitraryFilterLoad = ArbitraryFilterLoad FilterLoad
-    deriving (Eq, Show, Read)
-
-instance Arbitrary ArbitraryFilterLoad where
-    arbitrary = do
-        ArbitraryBloomFilter _ _ bf <- arbitrary
-        return $ ArbitraryFilterLoad $ FilterLoad bf
-
--- | Arbitrary FilterAdd
-data ArbitraryFilterAdd = ArbitraryFilterAdd FilterAdd
-    deriving (Eq, Show, Read)
-
-instance Arbitrary ArbitraryFilterAdd where
-    arbitrary = do
-        ArbitraryByteString bs <- arbitrary
-        return $ ArbitraryFilterAdd $ FilterAdd bs
