+++ Network/Haskoin/Transaction/Builder.hs
@@ -39,7 +39,7 @@ import Data.Aeson
 
 import Network.Haskoin.Util
 import Network.Haskoin.Crypto
-import Network.Haskoin.Protocol
+import Network.Haskoin.Node
 import Network.Haskoin.Script
 import Network.Haskoin.Transaction.Types
