+++ haskoin.cabal
@@ -33,7 +33,7 @@ library
     exposed-modules:   Network.Haskoin.Util,
                        Network.Haskoin.Stratum,
                        Network.Haskoin.Crypto,
-                       Network.Haskoin.Protocol,
+                       Network.Haskoin.Node,
                        Network.Haskoin.Script,
                        Network.Haskoin.Transaction,
                        Network.Haskoin.Block,
@@ -51,9 +51,9 @@ library
                        Network.Haskoin.Crypto.NormalizedKeys,
                        Network.Haskoin.Crypto.ECDSA,
                        Network.Haskoin.Crypto.Mnemonic,
-                       Network.Haskoin.Crypto.Bloom,
-                       Network.Haskoin.Protocol.Message,
-                       Network.Haskoin.Protocol.Types,
+                       Network.Haskoin.Node.Types,
+                       Network.Haskoin.Node.Message,
+                       Network.Haskoin.Node.Bloom,
                        Network.Haskoin.Script.Types, 
                        Network.Haskoin.Script.Parser, 
                        Network.Haskoin.Script.SigHash,
@@ -63,7 +63,7 @@ library
                        Network.Haskoin.Block.Types,
                        Network.Haskoin.Block.Merkle,
                        Network.Haskoin.Test.Crypto,
-                       Network.Haskoin.Test.Protocol,
+                       Network.Haskoin.Test.Node,
                        Network.Haskoin.Test.Message,
                        Network.Haskoin.Test.Script,
                        Network.Haskoin.Test.Transaction,
@@ -107,8 +107,8 @@ test-suite test-haskoin
                        Network.Haskoin.Crypto.Hash.Units,
                        Network.Haskoin.Crypto.Mnemonic.Tests,
                        Network.Haskoin.Crypto.Mnemonic.Units,
-                       Network.Haskoin.Crypto.Bloom.Units,
                        Network.Haskoin.Crypto.Units,
+                       Network.Haskoin.Node.Units,
                        Network.Haskoin.Script.Tests,
                        Network.Haskoin.Script.Units,
                        Network.Haskoin.Transaction.Tests,
