+++ tests/Network/Haskoin/Binary/Tests.hs
@@ -6,7 +6,7 @@ import Test.Framework.Providers.QuickCheck2 (testProperty)
 import Data.Binary (Binary)
 
 import Network.Haskoin.Test.Crypto
-import Network.Haskoin.Test.Protocol
+import Network.Haskoin.Test.Node
 import Network.Haskoin.Test.Message
 import Network.Haskoin.Test.Script
 import Network.Haskoin.Test.Transaction
